refactor(i18n): extract the locale logic into ssg-i18n (#588)
Locale negotiation, hreflang link building, URL-prefix strategy and
language-switcher markup were pure functions of their arguments living
inside a plugin. Nothing outside ssg could use them, and nothing inside
ssg could test them without a site.

They now live in `crates/ssg-i18n`, a leaf crate with no dependency on
ssg — no `Plugin`, no `SsgError`, no file I/O — alongside `ssg-a11y` and
`ssg-search`. What stayed behind is everything that walks or writes the
filesystem: locale detection, page collection, sitemap emission and the
`Plugin` impl that drives them. `src/plugins/i18n.rs` goes 3731 -> 3260
lines.

A mechanical move, and checked as one rather than asserted:

- The diff to `src/plugins/i18n.rs` is 472 deletions and 10 insertions,
  every insertion a re-export or its comment.
- Every deleted line was matched against the new crate. The 13 that did
  not match are each accounted for: 9 visibility widenings needed to
  cross the crate boundary, 3 doctest import paths, and one doc summary
  split for `clippy::too_long_first_doc_paragraph`.
- 9 of the 14 moved items were crate-private in ssg. They have to be
  `pub` in ssg-i18n, so they are re-exported `pub(crate)` — a blanket
  `pub use` would have silently widened ssg's public API.
  `tests/api_surface_i18n.rs` asserts the 5 previously-public items are
  still reachable; eight `compile_fail` doctests assert the private ones
  are not. Both halves were checked by mutation: swapping a public item
  into a `compile_fail` block makes it fail.

The crate's 12 unit tests were likewise mutation-checked — dropping the
empty-placeholder guard, truncating the element span, removing the
region-to-base-language fallback, collapsing SubDomain into SubPath,
inverting the root-locale check and reversing the q-sort each produce a
failure. The README's examples are wired in as doctests, so its
documented outputs cannot drift from the code.

`ssg-i18n` is added to the release workflow's publish order and to
`PUBLISHED_MEMBERS` in `tests/release_versions.rs`: ssg now depends on
it by path and version, so `cargo publish -p ssg` would fail without it,
and the gate that compares those two inventories caught the second one
being missing.

It is not yet on crates.io, and Trusted Publishing tokens cannot create
a crate — the trap that killed 0.0.57 through 0.0.61. `cargo publish -p
ssg-i18n` has to be run once by hand before v0.0.63 is tagged.
