+++ Cargo.toml
@@ -167,6 +167,7 @@ rayon = "1.12.0"
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.149"
 ssg-a11y = { path = "crates/ssg-a11y", version = "0.0.62" }
+ssg-i18n = { path = "crates/ssg-i18n", version = "0.0.62" }
 ssg-core = { path = "crates/ssg-core", version = "0.0.62" }
 ssg-rpc = { path = "crates/ssg-rpc", version = "0.0.62" }
 ssg-search = { path = "crates/ssg-search", version = "0.0.62" }
@@ -569,6 +570,7 @@ members = [
     "crates/ssg-search",
     "crates/ssg-wasm",
     "crates/ssg-heap-probe",
+    "crates/ssg-i18n",
 ]
 default-members = ["."]
