+++ README.md
@@ -220,8 +220,8 @@ graph TD
 
 | Metric | Value |
 | :--- | :--- |
-| **Source** | 143,000+ lines across 9 workspace crates (`ssg`, `ssg-core`, `ssg-a11y`, `ssg-search`, `ssg-mcp`, `ssg-rpc`, `ssg-rpc-macro`, `ssg-wasm`, and the unpublished `ssg-heap-probe`) |
-| **Test suite** | 3,668 unit tests (`cargo test --lib`) + 58 integration test targets |
+| **Source** | 143,000+ lines across 10 workspace crates (`ssg`, `ssg-core`, `ssg-a11y`, `ssg-i18n`, `ssg-search`, `ssg-mcp`, `ssg-rpc`, `ssg-rpc-macro`, `ssg-wasm`, and the unpublished `ssg-heap-probe`) |
+| **Test suite** | 3,668 unit tests (`cargo test --lib`) + 59 integration test targets |
 | **Coverage** | 98% region, 98% line, 98% function (CI-gated); measured 99.31 / 99.29 / 99.24 |
 | **Plugin pipeline** | 32 plugins, Rayon-parallelised |
 | **Audit gates** | 15 (WCAG 2.2 AAA, JSON-LD, hreflang, lang consistency, CSP+SRI, PQC TLS, HTML5, broken links, OG, markdown lint, perf budget, AI discovery, RSS/Atom, image opt, search index integrity) |
