+++ docs/features-coverage.md
@@ -23,7 +23,7 @@ name.
 | CSP build-time extraction | `csp` | [`examples/blog`](../examples/blog_example.rs), [`examples/landing`](../examples/landing_example.rs) | [`benches/all_pub_api.rs`](../benches/all_pub_api.rs) | [`tests/csp_preserve_attrs.rs`](../tests/csp_preserve_attrs.rs), [`tests/audit_gates.rs`](../tests/audit_gates.rs) |
 | Draft filtering | `drafts` | [`examples/blog`](../examples/blog_example.rs), [`examples/portfolio`](../examples/portfolio_example.rs) | [`benches/all_pub_api.rs`](../benches/all_pub_api.rs) | [`tests/plugins/drafts.rs`](../tests/plugins/drafts.rs) (via plugins/ submodule) |
 | Syntax highlighting | `highlight` | [`examples/basic`](../examples/basic_example.rs), [`examples/docs`](../examples/docs_example.rs) | [`benches/all_pub_api.rs`](../benches/all_pub_api.rs) | [`tests/regression.rs`](../tests/regression.rs) |
-| Internationalisation | `i18n` | [`examples/multilingual`](../examples/multilingual_example.rs), [`examples/multilingual_full`](../examples/multilingual_full_example.rs) (nested `content/<lang>/` tree; exercises `staticdatagen 0.0.10` recursive walk) | [`benches/all_pub_api.rs`](../benches/all_pub_api.rs) | [`tests/regression.rs`](../tests/regression.rs), [`tests/regression_user_site.rs::nested_locale_subdirectories_build_per_language`](../tests/regression_user_site.rs) (currently `#[ignore]`d pending the `staticdatagen 0.0.10` dep bump on `feat/v0.0.46`) |
+| Internationalisation | `i18n` | [`examples/multilingual`](../examples/multilingual_example.rs), [`examples/multilingual_full`](../examples/multilingual_full_example.rs) (nested `content/<lang>/` tree; exercises `staticdatagen 0.0.10` recursive walk) | [`benches/all_pub_api.rs`](../benches/all_pub_api.rs) | [`tests/regression.rs`](../tests/regression.rs), [`tests/regression_user_site.rs::nested_locale_subdirectories_build_per_language`](../tests/regression_user_site.rs) (currently `#[ignore]`d pending the `staticdatagen 0.0.10` dep bump on `feat/v0.0.46`) — the locale-negotiation, hreflang and URL-strategy logic now lives in the standalone [`crates/ssg-i18n`](../crates/ssg-i18n) crate, unit-tested in-crate (`cargo test -p ssg-i18n`); `src/plugins/i18n.rs` keeps the filesystem walking and the `Plugin` impl, covered by its own `#[cfg(test)] mod tests` (`cargo test --lib -- plugins_group::i18n`) |
 | Responsive image pipeline | `image_plugin` | [`examples/blog`](../examples/blog_example.rs), [`examples/portfolio`](../examples/portfolio_example.rs) | [`benches/avif_vs_webp.rs`](../benches/avif_vs_webp.rs) | [`tests/plugins/image_plugin.rs`](../tests/plugins/image_plugin.rs) |
 | Web-Components islands | `islands` | [`examples/docs`](../examples/docs_example.rs), [`examples/landing`](../examples/landing_example.rs) | [`benches/all_pub_api.rs`](../benches/all_pub_api.rs) | [`tests/regression.rs`](../tests/regression.rs) |
 | ISR manifest emission | `isr_manifest` | [`examples/isr`](../examples/isr_example.rs) | [`benches/all_pub_api.rs`](../benches/all_pub_api.rs) | [`tests/isr_manifest_shape.rs`](../tests/isr_manifest_shape.rs), [`tests/isr_edge_contract.rs`](../tests/isr_edge_contract.rs), [`tests/isr_backcompat.rs`](../tests/isr_backcompat.rs) |
